backend/requirements.txt
Flask
flask-SQLAlchemy
flask-Smorest
flask-migrate
python-dotenv
SQLAlchemy
Flask-Cors
gunicorn
psycopg2
